requirements.txt
websocket-client
aiohttp
